Addressing some more PR review comments
