+++ com.unity.netcode.gameobjects/Runtime/Messaging/DisconnectReasonMessage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d7742516058394f96999464f3ea32c71
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
